test: add shared fixtures in conftest.py
Consolidate duplicate test helpers (DummyApprovalManager, make_ctx, run)
into tests/conftest.py and update 5 test files to import from there.
This removes ~100 lines of duplicated boilerplate across the test suite.
