+++ tests/test_options.py
@@ -1,18 +1,10 @@
-import asyncio
 import datetime
 from enum import Enum
-from types import SimpleNamespace
-from typing import Any, cast
+from typing import Any
 
-from schwab.client import AsyncClient
 from schwab_mcp.tools import options
-from schwab_mcp.context import SchwabContext, SchwabServerContext
-from schwab_mcp.approvals import ApprovalDecision, ApprovalManager, ApprovalRequest
 
-
-class DummyApprovalManager(ApprovalManager):
-    async def require(self, request: ApprovalRequest) -> ApprovalDecision:  # noqa: ARG002
-        return ApprovalDecision.APPROVED
+from conftest import make_ctx, run
 
 
 class DummyOptionsClient:
@@ -31,22 +23,6 @@ class DummyOptionsClient:
 
     async def get_option_expiration_chain(self, *args, **kwargs):
         return None
-
-
-def run(coro):
-    return asyncio.run(coro)
-
-
-def make_ctx(client: Any) -> SchwabContext:
-    lifespan_context = SchwabServerContext(
-        client=cast(AsyncClient, client),
-        approval_manager=DummyApprovalManager(),
-    )
-    request_context = SimpleNamespace(lifespan_context=lifespan_context)
-    return SchwabContext.model_construct(
-        _request_context=cast(Any, request_context),
-        _fastmcp=None,
-    )
 
 
 def test_get_advanced_option_chain_parses_and_maps_parameters(monkeypatch):
